@@ -22,7 +22,7 @@
     "autoprefixer": "^10.4.16",
     "playwright": "^1.40.0",
     "postcss": "^8.4.32",
-    "svelte": "^4.2.7",
+    "svelte": "^5.0.0",
     "svelte-check": "^3.6.0",
     "tailwindcss": "^3.4.0",
     "tslib": "^2.4.1",
