@@ -1,17 +1,10 @@
-import { getCalendarData } from '$lib/services/calendar';
 import type { PageLoad } from './$types';
 
+// Don't fetch data on initial load - wait for courseid from dialog
 export const load: PageLoad = async () => {
-  try {
-    const calendarData = await getCalendarData();
-    return {
-      calendarData,
-      error: null
-    };
-  } catch (error) {
-    return {
-      calendarData: [],
-      error: error instanceof Error ? error.message : 'Failed to load calendar data'
-    };
-  }
+  return {
+    calendarData: [],
+    error: null,
+    courseid: null
+  };
 };
